+++ adminforth/spa/src/afcl/index.ts
@@ -7,3 +7,5 @@ export { default as Input } from './Input.vue';
 export { default as Tooltip } from './Tooltip.vue';
 export { default as LinkButton } from './LinkButton.vue';
 export { default as VerticalTabs } from './VerticalTabs.vue';
+export { default as Checkbox } from './Checkbox.vue';
+export { default as Dropzone } from './Dropzone.vue';