+++ adminforth/spa/src/components/ResourceListTable.vue
@@ -449,7 +449,6 @@ async function deleteRecord(row) {
           primaryKey: row._primaryKeyValue,
         }
       });
-      console.log('safsa', res)
       if (!res.error){
         emits('update:records', true)
         showSuccesTost('Record deleted successfully')